some test reorg
DRY test stuff, and have tests for the reference impl that compares with bao
