+++ src/iter.rs
@@ -8,7 +8,7 @@ use range_collections::{RangeSet2, RangeSetRef};
 use self_cell::self_cell;
 use smallvec::SmallVec;
 
-use crate::{blake3, BaoTree, BlockSize, ChunkNum, TreeNode};
+use crate::{split, BaoTree, BlockSize, ChunkNum, TreeNode};
 
 /// Extended node info.
 ///
@@ -505,82 +505,6 @@ impl<T: Default> Default for BaoChunk<T> {
     }
 }
 
-/// Reference implementation of the response iterator, using just the simple recursive
-/// implementation [crate::iter::select_nodes_rec].
-#[cfg(test)]
-pub(crate) fn partial_chunk_iter_reference(
-    tree: BaoTree,
-    ranges: &RangeSetRef<ChunkNum>,
-    min_full_level: u8,
-) -> Vec<BaoChunk<&RangeSetRef<ChunkNum>>> {
-    let mut res = Vec::new();
-    select_nodes_rec(
-        ChunkNum(0),
-        tree.size.to_usize(),
-        true,
-        ranges,
-        tree.block_size.to_u32(),
-        min_full_level as u32,
-        &mut |x| res.push(x),
-    );
-    res
-}
-
-/// Reference implementation of the response iterator, using just the simple recursive
-/// implementation [crate::iter::select_nodes_rec].
-#[cfg(test)]
-pub(crate) fn response_iter_reference(
-    tree: BaoTree,
-    ranges: &RangeSetRef<ChunkNum>,
-) -> Vec<BaoChunk> {
-    let mut res = Vec::new();
-    select_nodes_rec(
-        ChunkNum(0),
-        tree.size.to_usize(),
-        true,
-        ranges,
-        0,
-        tree.block_size.to_u32(),
-        &mut |x| res.push(x.without_ranges()),
-    );
-    res
-}
-
-/// A reference implementation of [PreOrderPartialChunkIterRef] that just uses the recursive
-/// implementation [crate::iter::select_nodes_rec].
-///
-/// This is not a proper iterator since it computes all elements at once, but it is useful for
-/// testing.
-#[cfg(test)]
-#[derive(Debug)]
-pub struct ReferencePreOrderPartialChunkIterRef<'a> {
-    iter: std::vec::IntoIter<BaoChunk<&'a RangeSetRef<ChunkNum>>>,
-    tree: BaoTree,
-}
-
-#[cfg(test)]
-impl<'a> ReferencePreOrderPartialChunkIterRef<'a> {
-    /// Create a new iterator over the tree.
-    pub fn new(tree: BaoTree, ranges: &'a RangeSetRef<ChunkNum>, min_full_level: u8) -> Self {
-        let iter = partial_chunk_iter_reference(tree, ranges, min_full_level).into_iter();
-        Self { iter, tree }
-    }
-
-    /// Return a reference to the underlying tree.
-    pub fn tree(&self) -> &BaoTree {
-        &self.tree
-    }
-}
-
-#[cfg(test)]
-impl<'a> Iterator for ReferencePreOrderPartialChunkIterRef<'a> {
-    type Item = BaoChunk<&'a RangeSetRef<ChunkNum>>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.iter.next()
-    }
-}
-
 /// Iterator over all nodes in a BaoTree in pre-order that overlap with a given chunk range.
 ///
 /// This is mostly used internally
@@ -819,198 +743,5 @@ impl Iterator for ResponseIter {
 
     fn next(&mut self) -> Option<Self::Item> {
         self.0.next()
-    }
-}
-
-/// Split and canonicalize a range set at a given chunk number
-///
-/// Compared to [RangeSetRef::split], this function will canonicalize the second range
-pub(crate) fn split(
-    ranges: &RangeSetRef<ChunkNum>,
-    mid: ChunkNum,
-) -> (&RangeSetRef<ChunkNum>, &RangeSetRef<ChunkNum>) {
-    let (mut a, mut b) = ranges.split(mid);
-    // todo: the canonicalization for a should be done in RangeSetRef::split.
-    // this is just a workaround until that bug is fixed.
-    if a.boundaries().last() == Some(&mid) {
-        a = RangeSetRef::new(&a.boundaries()[..a.boundaries().len() - 1]).unwrap();
-    }
-    if b.boundaries().len() == 1 && b.boundaries()[0] <= mid {
-        b = RangeSetRef::new(&[ChunkNum(0)]).unwrap();
-    }
-    (a, b)
-}
-
-/// Encode ranges relevant to a query from a slice and outboard to a buffer.
-///
-/// This will compute the root hash, so it will have to traverse the entire tree.
-/// The `ranges` parameter just controls which parts of the data are written.
-///
-/// Except for writing to a buffer, this is the same as [hash_subtree].
-/// The `min_level` parameter controls the minimum level that will be emitted as a leaf.
-/// Set this to 0 to disable chunk groups entirely.
-/// The `emit_data` parameter controls whether the data is written to the buffer.
-/// When setting this to false and setting query to `RangeSet::all()`, this can be used
-/// to write an outboard.
-///
-/// `res` will not contain the length prefix, so if you want a bao compatible format,
-/// you need to prepend it yourself.
-///
-/// This is used as a reference implementation in tests, but also to compute hashes
-/// below the chunk group size when creating responses for outboards with a chunk group
-/// size of >0.
-pub(crate) fn encode_selected_rec(
-    start_chunk: ChunkNum,
-    data: &[u8],
-    is_root: bool,
-    query: &RangeSetRef<ChunkNum>,
-    min_level: u32,
-    emit_data: bool,
-    res: &mut Vec<u8>,
-) -> blake3::Hash {
-    use blake3::guts::{ChunkState, CHUNK_LEN};
-    if data.len() <= CHUNK_LEN {
-        if emit_data && !query.is_empty() {
-            res.extend_from_slice(data);
-        }
-        let mut hasher = ChunkState::new(start_chunk.0);
-        hasher.update(data);
-        hasher.finalize(is_root)
-    } else {
-        let chunks = data.len() / CHUNK_LEN + (data.len() % CHUNK_LEN != 0) as usize;
-        let chunks = chunks.next_power_of_two();
-        let level = chunks.trailing_zeros() - 1;
-        let mid = chunks / 2;
-        let mid_bytes = mid * CHUNK_LEN;
-        let mid_chunk = start_chunk + (mid as u64);
-        let (l_ranges, r_ranges) = split(query, mid_chunk);
-        // for empty ranges, we don't want to emit anything.
-        // for full ranges where the level is below min_level, we want to emit
-        // just the data.
-        //
-        // todo: maybe call into blake3::guts::hash_subtree directly for this case? it would be faster.
-        let full = query.is_all();
-        let emit_parent = !query.is_empty() && (!full || level >= min_level);
-        let hash_offset = if emit_parent {
-            // make some room for the hashes
-            res.extend_from_slice(&[0xFFu8; 64]);
-            Some(res.len() - 64)
-        } else {
-            None
-        };
-        // recurse to the left and right to compute the hashes and emit data
-        let left = encode_selected_rec(
-            start_chunk,
-            &data[..mid_bytes],
-            false,
-            l_ranges,
-            min_level,
-            emit_data,
-            res,
-        );
-        let right = encode_selected_rec(
-            mid_chunk,
-            &data[mid_bytes..],
-            false,
-            r_ranges,
-            min_level,
-            emit_data,
-            res,
-        );
-        // backfill the hashes if needed
-        if let Some(o) = hash_offset {
-            res[o..o + 32].copy_from_slice(left.as_bytes());
-            res[o + 32..o + 64].copy_from_slice(right.as_bytes());
-        }
-        blake3::guts::parent_cv(&left, &right, is_root)
-    }
-}
-
-/// Select nodes relevant to a query
-///
-/// This is the receive side equivalent of [bao_encode_selected_recursive].
-/// It does not do any hashing, but just emits the nodes that are relevant to a query.
-///
-/// The tree is given as `start_chunk`, `size` and `is_root`.
-///
-/// To traverse an entire tree, use `0` for `start_chunk`, `tree.size` for `size` and
-/// `true` for `is_root`.
-///
-/// `tree_level` is the smallest level the iterator will emit. Set this to `tree.block_size`
-/// `min_full_level` is the smallest level that will be emitted as a leaf if it is fully
-/// within the query range.
-///
-/// To disable chunk groups entirely, just set both `tree_level` and `min_full_level` to 0.
-#[cfg(test)]
-pub(crate) fn select_nodes_rec<'a>(
-    start_chunk: ChunkNum,
-    size: usize,
-    is_root: bool,
-    ranges: &'a RangeSetRef<ChunkNum>,
-    tree_level: u32,
-    min_full_level: u32,
-    emit: &mut impl FnMut(BaoChunk<&'a RangeSetRef<ChunkNum>>),
-) {
-    if ranges.is_empty() {
-        return;
-    }
-    use blake3::guts::CHUNK_LEN;
-    if size <= CHUNK_LEN {
-        emit(BaoChunk::Leaf {
-            start_chunk,
-            size,
-            is_root,
-            ranges,
-        });
-    } else {
-        let chunks: usize = size / CHUNK_LEN + (size % CHUNK_LEN != 0) as usize;
-        let chunks = chunks.next_power_of_two();
-        // chunks is always a power of two, 2 for level 0
-        // so we must subtract 1 to get the level, and this is also safe
-        let level = chunks.trailing_zeros() - 1;
-        let full = ranges.is_all();
-        if (level < tree_level) || (full && level < min_full_level) {
-            // we are allowed to just emit the entire data as a leaf
-            emit(BaoChunk::Leaf {
-                start_chunk,
-                size,
-                is_root,
-                ranges,
-            });
-        } else {
-            // split in half and recurse
-            assert!(start_chunk.0 % 2 == 0);
-            let mid = chunks / 2;
-            let mid_bytes = mid * CHUNK_LEN;
-            let mid_chunk = start_chunk + (mid as u64);
-            let (l_ranges, r_ranges) = split(ranges, mid_chunk);
-            let node = TreeNode::from_start_chunk_and_level(start_chunk, BlockSize(level as u8));
-            emit(BaoChunk::Parent {
-                node,
-                is_root,
-                left: !l_ranges.is_empty(),
-                right: !r_ranges.is_empty(),
-                ranges,
-            });
-            // recurse to the left and right to compute the hashes and emit data
-            select_nodes_rec(
-                start_chunk,
-                mid_bytes,
-                false,
-                l_ranges,
-                tree_level,
-                min_full_level,
-                emit,
-            );
-            select_nodes_rec(
-                mid_chunk,
-                size - mid_bytes,
-                false,
-                r_ranges,
-                tree_level,
-                min_full_level,
-                emit,
-            );
-        }
     }
 }
